Use actual data for histogram
