+++ apps/cde-ui/src/assets/color_mapping.csv
@@ -1,0 +1,46 @@
+cell_id,cell_type,cell_color
+1,Antigen Presenting Cell,"[255, 127, 14]"
+2,B Cell,"[44, 160, 44]"
+3,Basal Epithelial Cell,"[214, 39, 40]"
+4,CD14+ Classical Monocyte,"[148, 103, 189]"
+5,CD1c+ Dendritic Cell,"[140, 86, 75]"
+6,CD4+ T Cell,"[227, 119, 194]"
+7,CD8+ T Cell,"[127, 127, 127]"
+8,Endothelial Cell,"[188, 189, 34]"
+9,Epithelial,"[23, 190, 207]"
+10,Immune Cell,"[174, 199, 232]"
+11,Lymphatic Endothelial Cell (and some immune cells),"[255, 187, 120]"
+12,Macrophage,"[152, 223, 138]"
+13,Macrophages,"[255, 152, 150]"
+14,Mast Cell,"[197, 176, 213]"
+15,Megakaryocyte and Platelet,"[196, 156, 148]"
+16,Neutrophil,"[247, 182, 210]"
+17,Proliferating Cell,"[199, 199, 199]"
+18,Smooth Muscle Cell,"[219, 219, 141]"
+19,T Cell ,"[158, 218, 229]"
+20,T Regulatory Cell,"[255, 127, 14]"
+21,Type 1 Alveolar Epithelial Cell,"[44, 160, 44]"
+22,Type 2 Alveolar Epithelial Cell,"[31, 119, 180]"
+23,Blood endothelial,"[255, 0, 0]"
+24,Antigen Presenting Cell_link,"[255, 127, 14]"
+25,B Cell_link,"[44, 160, 44]"
+26,Basal Epithelial Cell_link,"[214, 39, 40]"
+27,CD14+ Classical Monocyte_link,"[148, 103, 189]"
+28,CD1c+ Dendritic Cell_link,"[140, 86, 75]"
+29,CD4+ T Cell_link,"[227, 119, 194]"
+30,CD8+ T Cell_link,"[127, 127, 127]"
+31,Endothelial Cell_link,"[188, 189, 34]"
+32,Epithelial_link,"[23, 190, 207]"
+33,Immune Cell_link,"[174, 199, 232]"
+34,Lymphatic Endothelial Cell (and some immune cells)_link,"[255, 187, 120]"
+35,Macrophage_link,"[152, 223, 138]"
+36,Macrophages_link,"[255, 152, 150]"
+37,Mast Cell_link,"[197, 176, 213]"
+38,Megakaryocyte and Platelet_link,"[196, 156, 148]"
+39,Neutrophil_link,"[247, 182, 210]"
+40,Proliferating Cell_link,"[199, 199, 199]"
+41,Smooth Muscle Cell_link,"[219, 219, 141]"
+42,T Cell _link,"[158, 218, 229]"
+43,T Regulatory Cell_link,"[255, 127, 14]"
+44,Type 1 Alveolar Epithelial Cell_link,"[44, 160, 44]"
+45,Type 2 Alveolar Epithelial Cell_link,"[31, 119, 180]"
