+++ libs/cde-visualization/src/lib/components/histogram/histogram.component.scss
@@ -1,0 +1,162 @@
+:host {
+  ::ng-deep mat-expansion-panel {
+    --mat-expansion-container-shape: 0.5rem;
+    --mat-expansion-container-background-color: white;
+    letter-spacing: 0.005em;
+    font-family: 'Metropolis';
+
+    &.mat-expanded {
+      overflow: visible;
+
+      .mat-expansion-panel-content {
+        opacity: 1;
+        transition: opacity 255ms;
+      }
+    }
+
+    .mat-expansion-panel-content {
+      opacity: 0;
+      transition-delay: 255ms !important;
+    }
+
+    mat-expansion-panel-header {
+      padding: 0.75rem;
+
+      .mat-content {
+        margin-right: 0;
+      }
+
+      .title {
+        font-size: 0.9375rem;
+        gap: 0.25rem;
+        user-select: none;
+
+        mat-icon {
+          vertical-align: middle;
+        }
+      }
+    }
+
+    .mat-expansion-panel-body {
+      padding: 0 0.75rem 0.75rem;
+    }
+  }
+
+  .download {
+    height: 1.75rem;
+    --mdc-filled-button-container-color: #d5dbe3;
+    --mdc-filled-button-label-text-color: #4b4b5e;
+    --mat-filled-button-state-layer-color: #4b4b5e;
+    --mat-filled-button-ripple-color: rgba(75, 75, 94, 0.1);
+
+    &:first-of-type {
+      margin-right: 1rem;
+    }
+  }
+
+  .histogram-container {
+    display: flex;
+    justify-content: space-between;
+    gap: 2rem;
+    max-height: 18.5rem;
+
+    .visualization {
+      width: calc(100vw - 39rem);
+    }
+
+    .legend {
+      width: 15.5rem;
+      overflow-y: auto;
+      font-size: 12px;
+      font-weight: 400;
+
+      .legend-entry {
+        display: flex;
+        align-items: center;
+        text-wrap: nowrap;
+      }
+
+      .color-icon {
+        height: 1.125rem;
+        width: 1.125rem;
+        min-width: 1.125rem;
+        border-radius: 2rem;
+        margin: 0.25rem;
+        margin-right: 0.5rem;
+      }
+
+      .anchor {
+        --mdc-checkbox-unselected-icon-color: #201e3d;
+        border-radius: 50%;
+        color: white;
+        vertical-align: middle;
+        height: 1.125rem;
+        width: 1.125rem;
+        font-size: 1rem;
+        text-align: center;
+        align-content: center;
+        background-color: var(--mdc-checkbox-unselected-icon-color);
+        margin-right: 0.5rem;
+      }
+    }
+  }
+
+  // Color picker styling
+  ::ng-deep .color-picker {
+    border-radius: 0.5rem;
+    border: none;
+    box-shadow: 0px 5px 16px 0px #201e3d3d;
+
+    .hex-text {
+      display: flex !important;
+      flex-direction: row-reverse;
+      padding: 0 0.5rem;
+
+      .box {
+        width: 100%;
+        padding: 1.25rem 0;
+        align-items: center;
+
+        input {
+          border: none;
+          box-shadow: 0px 5px 16px 0px #201e3d3d;
+          border-radius: 1rem;
+          padding: 0.5625rem 0;
+          height: 2.5rem;
+        }
+      }
+    }
+
+    .arrow.arrow-right,
+    .type-policy {
+      display: none;
+    }
+
+    .button-area {
+      display: flex;
+      flex-direction: row-reverse;
+    }
+    .cursor {
+      height: 0.5rem;
+      width: 0.5rem;
+      border-color: white;
+    }
+
+    .selected-color {
+      border: none;
+    }
+
+    .saturation-lightness {
+      width: 86%;
+      margin: 1rem;
+    }
+  }
+
+  ::ng-deep .hue-alpha {
+    .hue,
+    .alpha {
+      border-radius: 0.25rem;
+      height: 0.5rem;
+    }
+  }
+}
