--- libs/cde-visualization/src/lib/histogram/histogram.component.html
@@ -1,51 +1,0 @@
-<mat-accordion>
-  <mat-expansion-panel hideToggle (opened)="panelOpen = true" (closed)="panelOpen = false" expanded>
-    <mat-expansion-panel-header>
-      <mat-panel-title class="mat-caption title">
-        <mat-icon class="material-symbols-rounded" *ngIf="!panelOpen">add</mat-icon>
-        <mat-icon class="material-symbols-rounded" *ngIf="panelOpen">remove</mat-icon>
-        Histogram
-        <mat-icon class="material-symbols-rounded">info</mat-icon>
-      </mat-panel-title>
-      <!-- <div class="download-buttons"> -->
-      <button mat-flat-button color="primary" class="mat-caption download" (click)="download($event, 'svg')">
-        <mat-icon class="material-symbols-rounded">download</mat-icon>
-        SVG
-      </button>
-      <button mat-flat-button color="primary" class="mat-caption download" (click)="download($event, 'png')">
-        <mat-icon class="material-symbols-rounded">download</mat-icon>
-        PNG
-      </button>
-      <!-- </div> -->
-    </mat-expansion-panel-header>
-    <div class="histogram-container">
-      <div class="visualization" #vis></div>
-      <div class="legend">
-        <div class="legend-entry">
-          <div class="color-icon" [style.background]="'black'"></div>
-          All Cells
-        </div>
-        <div class="legend-entry" *ngFor="let color of colorMap">
-          <!-- Icon with color picker -->
-          <!-- <div
-            class="color-icon"
-            (colorPickerOpen)="setCurrentColor(color.cell_color)"
-            [style.background]="toRGB(color.cell_color)"
-            [(colorPicker)]="currentColor"
-            [cpOutputFormat]="'hex'"
-            [cpOKButton]="true"
-            [cpCancelButton]="true"
-            [cpPosition]="'top-left'"
-            [cpOKButtonClass]="'accept-button'"
-            [cpOKButtonText]="'Accept'"
-            [cpCancelButtonClass]="'cancel-button'"
-            [cpAlphaChannel]="'always'"
-          ></div> -->
-          <span class="color-icon" [style.background]="toRGB(color.cell_color)"></span>
-          <mat-icon *ngIf="color.anchor" class="material-symbols-rounded anchor"> anchor </mat-icon>
-          {{ color.cell_type }}
-        </div>
-      </div>
-    </div>
-  </mat-expansion-panel>
-</mat-accordion>
