--- libs/cde-visualization/src/lib/histogram/histogram.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { HistogramComponent } from './histogram.component';
-
-describe('HistogramComponent', () => {
-  let component: HistogramComponent;
-  let fixture: ComponentFixture<HistogramComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [HistogramComponent],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(HistogramComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
